add expression parser with AST generation and simple sequential evaluator

+++ src/main/scala/parcalc/JsonSupport.scala
@@ -4,7 +4,7 @@ import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport
 import spray.json.DefaultJsonProtocol
 
 case class EvaluationRequest(expression: String)
-case class EvaluationResponse(result: Int)
+case class EvaluationResponse(result: Double)
 
 trait JsonSupport extends SprayJsonSupport with DefaultJsonProtocol {
   implicit val evaluationRequestFormat = jsonFormat1(EvaluationRequest)
